refactor: bottomSheet 커스텀 훅 리팩토링
- useDraggable 커스텀 훅 제작
- translateY는 CSS에서 쓰는 단어라서 React Props로는 권장하지 않는다고 하여 변경
